requirements.txt
python
django-cors-headers
djoser
djangorestframework==3.11.2
Django==3.2.10
psycopg2